part3/webapp/live/pypi_web_mongodb_live/requirements-dev.txt
-r requirements.txt

waitress==1.1.0
Mako==1.0.7
MarkupSafe==1.0
Pygments==2.2.0
pyramid-debugtoolbar==4.4
pyramid-mako==1.0.2
WebTest==2.0.29 
atomicwrites==1.1.5 
attrs==18.1.0 
beautifulsoup4==4.6.0 
coverage==4.5.1 
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.1 
pytest-cov==2.5.1
